refactor: move third party auth logic into a helper
